adding upgrade tests to main test file and removing temporary test files
